requirements.txt
# requirements.txt
google-generativeai==0.3.2
faiss-cpu==1.7.4
psycopg2-binary==2.9.7
PyPDF2==3.0.1
python-docx==0.8.11
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
tiktoken==0.5.1
numpy==1.24.3